requirements.txt
reportlab
arabic-reshaper
python-bidi
pywin32
Pillow
pandas
openpyxl
pdfplumber
